+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.9.14 - Unreleased
 
+### Added
+* New function: `exactly_1channel`.
+
 
 ## Version 0.9.13 - Released 2023-03-04