+++ CHANGELOG.md
@@ -4,7 +4,16 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.9.13 - Unreleased
+## Version 0.9.14 - Unreleased
+
+### Added
+* New function: `exactly_1channel`.
+
+### Fixed
+* kwimage.imresize now works when the input image has dsize=(1, 1)
+
+
+## Version 0.9.13 - Released 2023-03-04
 
 ### Changed
 * Small speedups